+++ warp/examples/fem/example_diffusion.py
@@ -17,20 +17,10 @@
 ###########################################################################
 
 import warp as wp
+import warp.examples.fem.utils as fem_example_utils
 import warp.fem as fem
 from warp.fem.utils import array_axpy
 from warp.sparse import bsr_axpy
-
-# Import example utilities
-# Make sure that works both when imported as module and run as standalone file
-try:
-    from .bsr_utils import bsr_cg
-    from .mesh_utils import gen_quadmesh, gen_trimesh
-    from .plot_utils import Plot
-except ImportError:
-    from bsr_utils import bsr_cg
-    from mesh_utils import gen_quadmesh, gen_trimesh
-    from plot_utils import Plot
 
 
 @fem.integrand
@@ -92,10 +82,10 @@ class Example:
 
         # Grid or triangle mesh geometry
         if mesh == "tri":
-            positions, tri_vidx = gen_trimesh(res=wp.vec2i(resolution))
+            positions, tri_vidx = fem_example_utils.gen_trimesh(res=wp.vec2i(resolution))
             self._geo = fem.Trimesh2D(tri_vertex_indices=tri_vidx, positions=positions)
         elif mesh == "quad":
-            positions, quad_vidx = gen_quadmesh(res=wp.vec2i(resolution))
+            positions, quad_vidx = fem_example_utils.gen_quadmesh(res=wp.vec2i(resolution))
             self._geo = fem.Quadmesh2D(quad_vertex_indices=quad_vidx, positions=positions)
         else:
             self._geo = fem.Grid2D(res=wp.vec2i(resolution))
@@ -107,7 +97,7 @@ class Example:
         # Scalar field over our function space
         self._scalar_field = self._scalar_space.make_field()
 
-        self.renderer = Plot()
+        self.renderer = fem_example_utils.Plot()
 
     def step(self):
         geo = self._geo
@@ -145,7 +135,7 @@ class Example:
 
         # Solve linear system using Conjugate Gradient
         x = wp.zeros_like(rhs)
-        bsr_cg(matrix, b=rhs, x=x, quiet=self._quiet)
+        fem_example_utils.bsr_cg(matrix, b=rhs, x=x, quiet=self._quiet)
 
         # Assign system result to our discrete field
         self._scalar_field.dof_values = x